+++ PokemonProject/src/attacks/electric/Spark.java
@@ -15,7 +15,7 @@ public class Spark extends PokemonAttack {
     }
 
     @Override
-    public double attack(User user, User enemyUser) {
+    public int attack(User user, User enemyUser) {
         // get only those pokemons that are currently in the battle
         Pokemon userPokemon = user.getCurrentPokemonForBattle();
         Pokemon enemyPokemon = enemyUser.getCurrentPokemonForBattle();
@@ -29,7 +29,7 @@ public class Spark extends PokemonAttack {
             finalAttackPower = 0.7 * this.attackPower;
         }
 
-        enemyPokemon.setHp(enemyPokemon.getHp() - finalAttackPower);
+        enemyPokemon.setHp((int) (enemyPokemon.getHp() - finalAttackPower));
 
         if (critChance == 0) {
             System.out.println(userPokemon.getName() + " missed the target!");
