+++ PokemonProject/src/attacks/normal/Leer.java
@@ -14,7 +14,7 @@ public class Leer extends PokemonAttack {
     }
 
     @Override
-    public double attack(User user, User enemyUser) {
+    public int attack(User user, User enemyUser) {
         // get only those pokemons that are currently in the battle
         Pokemon userPokemon = user.getCurrentPokemonForBattle();
         Pokemon enemyPokemon = enemyUser.getCurrentPokemonForBattle();
@@ -22,7 +22,7 @@ public class Leer extends PokemonAttack {
         double lowerEnemyPokemonDefence = 15;
 
         if (enemyPokemon.getDefencePoints() - lowerEnemyPokemonDefence >= 0) {
-            enemyPokemon.setDefencePoints(enemyPokemon.getDefencePoints() - lowerEnemyPokemonDefence);
+            enemyPokemon.setDefencePoints((int) (enemyPokemon.getDefencePoints() - lowerEnemyPokemonDefence));
         } else {
             enemyPokemon.setDefencePoints(0);
         }
